WORKING!!
The VarsFragment List can be changed and updates the adapter after you
close the activity. I didn't implement it on the FunFragment List yet as
the function items may not necessarily have a value to change. Still
needs documentation. Did not upload it to the dev branch as there was an
extra folder with other stuff in it that I did not want to erase. Let me
know if anyone finds any bugs while running it.