VarsFragment List Working
Able to update and pass variables to and from the VarsFragment List, the
ChangeWindowActivity, and the TelemetryApp class. Added a statement that
only allows the "editable" items to be changed. Working on
documentation. Did not apply it to the FunctionsFragment since I wasn't
sure if the same window would be applied.